requirements.txt
fastapi
uvicorn
pandas
mysql-connector-python
openpyxl
